+++ GovUkDesignSystemComponents/SubComponents/SummaryListRowKey.cshtml
@@ -1,12 +1,6 @@
 ﻿@model GovUkDesignSystem.GovUkDesignSystemComponents.SummaryListRowKey
 
-@{
-    string classNames =
-        $"govuk-summary-list__key "
-        + $"{(Model == null || String.IsNullOrEmpty(Model.Classes) ? string.Empty : Model.Classes)}";
-}
-
-<dt class="@classNames">
+<dt class="govuk-summary-list__key @Model?.Classes">
     @{
         if (Model != null)
         {
