+++ GovUkDesignSystemComponents/SummaryList.cshtml
@@ -1,12 +1,6 @@
 ﻿@model GovUkDesignSystem.GovUkDesignSystemComponents.SummaryListViewModel
 
-@{
-    string classNames =
-        $"govuk-summary-list "
-        + $"{(Model == null || String.IsNullOrEmpty(Model.Classes) ? string.Empty : Model.Classes)}";
-}
-
-<dl class="@classNames" @(Html.Raw(Model.Attributes != null ? string.Join(" ", Model.Attributes.Select(a => a.Key + "=\"" + a.Value + "\"")) : ""))>
+<dl class="govuk-summary-list @Model?.Classes" @(Html.Raw(Model.Attributes != null ? string.Join(" ", Model.Attributes.Select(a => a.Key + "=\"" + a.Value + "\"")) : ""))>
     @{
         if (Model.Rows != null && Model.Rows.Count > 0)
         {
